acts_as_solr and vlad the deployer.
solr.rake:
Use of a pipe instead of the command exec that was creating a subshell and never gave the prompt back.
Even the system command was not working after the command is forked.

Using IO.popen enabled me to bypass that and still get the pid of the solr process for the stop task.

It is now possible to start / stop the solr server remotely using rake:
rake staging vlad:solr:start
rake staging vlad:solr:stop

Feel free to patch your solr rakefile!